requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
binance-futures-connector==4.0.0
ccxt==4.0.80
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
dateparser==1.1.8
Django==4.2.4
frozenlist==1.4.0
gevent==23.7.0
greenlet==2.0.2
idna==3.4
kombu==5.3.2
multidict==6.0.4
mysqlclient==2.2.0
numpy==1.25.2
pandas==2.1.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycares==4.3.0
pycparser==2.21
pycryptodome==3.18.0
python-binance==1.0.19
python-dateutil==2.8.2
pyti==0.2.2
pytz==2023.3
redis==5.0.0
regex==2023.8.8
requests==2.31.0
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
ta==0.10.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.2
websockets==11.0.3
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
